backend/requirements.txt
# SmartFaceTrack Backend Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Data Processing & Validation
pydantic[email]==2.5.0
numpy==1.24.3

# Image Processing & Face Recognition
pillow==10.1.0
deepface==0.0.79
tensorflow==2.15.0
keras==2.15.0

# Environment & Utilities
python-dotenv==1.0.0

# Dependency resolution lock
typing-extensions==4.8.0
